refactor: db moved to separate file
