+++ src/db/db.ts
@@ -1,0 +1,4 @@
+export const users = [];
+export const tracks = [];
+export const artists = [];
+export const albums = [];
