+++ src/user/user.service.ts
@@ -3,8 +3,7 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { v4 as uuidv4 } from 'uuid';
 import { validate as uuidValidate } from 'uuid';
-
-const users = [];
+import { users } from 'src/db/db';
 
 @Injectable()
 export class UserService {
